MSBI Online Training
Techverze is the professional teaching platform provide MSBI online training, where Microsoft Business Intelligence is a power full suite and an ETL (Extract Transform and Load) tool helps to find solutions for BI and data mining queries. Each module is roofed with distinctive case study enable students to master the analytical strategies with MSBI and gain solid foundation with hands-on.